implement videostreambase
