+++ com.unity.template.renderstreaming-hd/Assets/Scripts/VideoStreamBase.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 8bcc82901c3f48a88d3408251afa3365
+timeCreated: 1606988906
